pyprog: Add BusPirate support
Add support for programming via BusPirate using the pyBusPirateLite
library.

The backend is negotiated automatically, BusPirate silently fallen back
to if the smbus backend doesn't work.

The --backend option allows explicitly choosing the backend, which is
useful for getting a traceback in case of a failure instead of the
silent fallback.

  $ time python pyprog.py -w sad-and-broken.bin
  Can't use SMBUS backend (use --backend=smbus to force it anyway), attempting BusPirate...
  Using BusPirate backend (assuming --backend=buspirate)
  JEDEC ID: 0x856013
  Manufacturer PUYA
  Chip: P25Q40
  Size: 512KB
  Setup chip commands for Winbond...
  Flash status register: 0x0
  Erasing...
  done
  Flashing sad-and-broken.bin
  Will write 512Kb
  Writting addr 6800
  ...